+++ WATA.LIS/WATA.LIS.Core/Model/VISION/VISON_Model.cs
@@ -11,7 +11,7 @@ namespace WATA.LIS.Core.Model.VISION
     {
         public float area { get; set; }
         public float width { get; set; }
-        public float height { get; set; }
+        public int height { get; set; }
         public string qr { get; set; }
         public string status { get; set; }
